requirements.txt
pymupdf==1.23.26
pdfplumber==0.11.4
tqdm==4.66.1
sentence-transformers>=2.7.0
chromadb>=0.5.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
